move visdata json parsing into a worker that can be linted and typechecked
