+++ .eslintrc.js
@@ -10,7 +10,7 @@ module.exports = {
     },
     parser: "@typescript-eslint/parser",
     parserOptions: {
-        project: "./tsconfig.json",
+        project: ["./tsconfig.json", "./src/visGeometry/workers/tsconfig.json"],
     },
     plugins: ["@typescript-eslint", "react"],
     rules: {
