+++ package.json
@@ -15,7 +15,7 @@
         "typeCheck": "tsc -p tsconfig.json --noEmit",
         "all-checks": "npm-run-all lint test typeCheck",
         "generateTypes": "tsc -p tsconfig.base.json --emitDeclarationOnly",
-        "transpileES": "cross-env TS_NODE_PROJECT=tsconfig.es.json BABEL_ENV=es babel src --out-dir es --extensions .ts,.tsx,.js,.jsx --ignore src/**/*.test.ts",
+        "transpileES": "cross-env TS_NODE_PROJECT=tsconfig.es.json BABEL_ENV=es babel src --out-dir es --extensions .ts,.tsx,.js,.jsx --ignore 'src/test'",
         "processCSS": "postcss src/viewport/index.css --output style/style.css",
         "build": "npm-run-all generateTypes transpileES processCSS",
         "copy-build": "copyfiles 'es/**/*' style/* 'type-declarations/**/*' package.json README.md dev-build",
