+++ src/visGeometry/workers/visDataWorker.ts
@@ -1,0 +1,16 @@
+import { parseVisDataMessage } from "../../simularium/VisDataParse";
+
+self.addEventListener(
+    "message",
+    (e: MessageEvent) => {
+        const visDataMsg = e.data;
+        const { frameDataArray, parsedAgentDataArray } =
+            parseVisDataMessage(visDataMsg);
+
+        postMessage({
+            frameDataArray,
+            parsedAgentDataArray,
+        });
+    },
+    false
+);
